# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for IPL Backend Service with menu management",
        "title": "IPL Backend Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/billings/bulk-monthly": {
            "post": {
                "description": "Create monthly billings for specified user IDs or all penghuni users if user_ids is empty. Requires auth-token cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billings"
                ],
                "summary": "Create bulk monthly billings",
                "parameters": [
                    {
                        "description": "Bulk billing request with month and year",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.BulkBillingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bulk billing creation result",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_internal_service.BulkBillingResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/menus/user/{user_id}": {
            "get": {
                "description": "Get list of menus accessible by a specific user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menus"
                ],
                "summary": "Get menus by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menus retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ipl-be-svc_internal_models_response.MenuResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "No menus found",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/billing/{id}/link": {
            "post": {
                "description": "Create a DOKU payment link for a billing record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create payment link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Billing ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link created successfully",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_internal_service.PaymentLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid billing ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Billing not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/users/penghuni": {
            "get": {
                "description": "Get list of all users with role type \"penghuni\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all penghuni users",
                "responses": {
                    "200": {
                        "description": "Penghuni users retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/ipl-be-svc_internal_models_response.PenghuniUserResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/profile/{user_id}": {
            "get": {
                "description": "Get detailed user information by profile ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user detail by profile ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detail retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handler.UserDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ipl-be-svc_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_handler.BulkBillingRequest": {
            "type": "object",
            "required": [
                "month",
                "year"
            ],
            "properties": {
                "month": {
                    "description": "Month 1-12",
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1
                },
                "user_ids": {
                    "description": "Empty means all penghuni users",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "year": {
                    "description": "Reasonable year range",
                    "type": "integer",
                    "maximum": 2100,
                    "minimum": 2020
                }
            }
        },
        "internal_handler.UserDetailResponse": {
            "type": "object",
            "properties": {
                "document_id": {
                    "type": "string",
                    "example": "abc123def456"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nama_penghuni": {
                    "type": "string",
                    "example": "John Doe"
                },
                "no_hp": {
                    "type": "string",
                    "example": "+6281234567890"
                },
                "no_telp": {
                    "type": "string",
                    "example": "021-12345678"
                },
                "role_id": {
                    "type": "integer",
                    "example": 1
                },
                "role_name": {
                    "type": "string",
                    "example": "Administrator"
                },
                "role_type": {
                    "type": "string",
                    "example": "admin"
                },
                "user_id": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "ipl-be-svc_internal_models_response.MenuResponse": {
            "type": "object",
            "properties": {
                "document_id": {
                    "type": "string",
                    "example": "mo5qqs8ezbruui07t91p6da8"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "kode_menu": {
                    "type": "string",
                    "example": "master-data"
                },
                "nama_menu": {
                    "type": "string",
                    "example": "Master Data"
                },
                "published_at": {
                    "type": "string",
                    "example": "2025-10-23T15:16:28.206Z"
                },
                "urutan_menu": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "ipl-be-svc_internal_models_response.PenghuniUserResponse": {
            "type": "object",
            "properties": {
                "document_id": {
                    "type": "string",
                    "example": "abc123def456"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nama_penghuni": {
                    "type": "string",
                    "example": "John Doe"
                },
                "no_hp": {
                    "type": "string",
                    "example": "+6281234567890"
                },
                "no_telp": {
                    "type": "string",
                    "example": "021-12345678"
                },
                "role_id": {
                    "type": "integer",
                    "example": 5
                },
                "role_name": {
                    "type": "string",
                    "example": "Penghuni"
                },
                "role_type": {
                    "type": "string",
                    "example": "penghuni"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "ipl-be-svc_internal_service.BulkBillingResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed_count": {
                    "type": "integer"
                },
                "success_count": {
                    "type": "integer"
                },
                "total_billings": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                }
            }
        },
        "ipl-be-svc_internal_service.PaymentLinkResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "billing_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "payment_url": {
                    "type": "string"
                }
            }
        },
        "ipl-be-svc_pkg_utils.APIResponse": {
            "description": "Standard API response structure",
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}